requirements.txt
mysql-connector-python
python-dotenv
markdownify
unidecode
pyyaml
liquidpy
pytz
Pillow
markdown
beautifulsoup4
tqdm
requests
csscompressor
feedgen
htmlmin
boto3
